Adds functionality to Add to Cart button
Adds addCartClick to check if the product amount is more than 0 else
the button does nothing then checks if item ID is already in cartData
to see if the item has been added to cart and if so changes the amount
contained within the object. If it is the first time an item is added
to cart addCartData. cartData, modifyCartData, and addCartData are all
taken from the Navbar module in order for the data within cart to
persist through all pages as Navbar is the parent page.