Adds ability for checkout buttons to change cart
Adds functionality to addClick and subtractClick to modify cart on
button press in checkout. Changes subtractClick to trigger
removeFromCartData on checking productAmount is 1 and moves setTotalPrice
to the top of the function to trigger regardless of the number in
productAmount. Changes price variable to be totalPrice changed into a
numerical value and changes all values in the price field to be set at
2 decimal places.